[#1]feat: 데미지 계산 공식 작성
- 계산 공식 작성
- 계산 공식 TDD 완료
- 일부 버프를 enemyBuff 리스트에 추가되도록 변경
- stats의 자료형 변경 int -> double
- 주석 일괄 추가
- 속성 정보 테이블에 추가

+++ src/main/java/com/github/syun9274/hsr_damage_calculator/calculator/component/BaseDmg.java
@@ -30,7 +30,7 @@ public class BaseDmg {
      */
     public double getBaseDmg(double skillMultiplier,
                              double extraMultiplier,
-                             int scalingAttribute,
+                             double scalingAttribute,
                              int extraDmg) {
         return (skillMultiplier + extraMultiplier) * scalingAttribute + extraDmg;
     }
